Initial server implementation: - changes precison to float32 - removes ebiten dependency from simulation - Adds json and csv rule visualization - Adds rule matrix render

+++ pkg/particlelifesim/particle/particle.go
@@ -1,13 +1,13 @@
 package particle
 
 type Particle struct {
-	x  float64
-	y  float64
-	Vx float64
-	Vy float64
+	x  float32
+	y  float32
+	Vx float32
+	Vy float32
 }
 
-func NewParticle(x, y float64) *Particle {
+func NewParticle(x, y float32) *Particle {
 	p := new(Particle)
 	p.SetX(x)
 	p.SetY(y)
@@ -16,18 +16,18 @@ func NewParticle(x, y float64) *Particle {
 }
 
 // The values are normalized before storage
-func (p *Particle) SetX(x float64) {
+func (p *Particle) SetX(x float32) {
 	p.x = x
 }
 
-func (p *Particle) SetY(y float64) {
+func (p *Particle) SetY(y float32) {
 	p.y = y
 }
 
-func (p *Particle) GetX() float64 {
+func (p *Particle) GetX() float32 {
 	return p.x
 }
 
-func (p *Particle) GetY() float64 {
+func (p *Particle) GetY() float32 {
 	return p.y
 }
